Moved to decimal field

+++ mimer/quiz/models.py
@@ -4,7 +4,7 @@ from django.contrib.auth.models import AbstractUser
 # Create your models here.
 
 class QuizUser(AbstractUser):
-    overall_score = models.SmallIntegerField(default=0)
+    overall_score = models.DecimalField(default=0, decimal_places=2, max_digits=100)
     sana = models.BooleanField(default=False)
 
 class IntegerRangeField(models.IntegerField):
